+++ include/unity/scopes/internal/ScopeMetadataImpl.h
@@ -59,7 +59,8 @@ public:
     bool location_data_needed() const;                           // optional (default: false)
     std::vector<std::string> child_scope_ids() const;            // optional (default: empty array)
     int version() const;                                         // optional (default: 0)
-    std::vector<std::string> tags() const;                       // optional (default: empty array)
+    std::vector<std::string> keywords() const;                   // optional (default: empty array)
+    bool is_aggregator() const;                                  // optional (default: false)
 
     void set_scope_id(std::string const& scope_id);
     void set_proxy(ScopeProxy const& proxy);
@@ -78,7 +79,8 @@ public:
     void set_location_data_needed(bool location_data_needed);
     void set_child_scope_ids(std::vector<std::string> const& ids);
     void set_version(int v);
-    void set_tags(std::vector<std::string> const& tags);
+    void set_keywords(std::vector<std::string> const& keywords);
+    void set_is_aggregator(bool is_aggregator);
 
     VariantMap serialize() const;
     void deserialize(VariantMap const& var);
@@ -105,7 +107,8 @@ private:
     std::unique_ptr<bool> location_data_needed_;          // Optional, hence a pointer
     std::vector<std::string> child_scope_ids_;
     int version_;
-    std::vector<std::string> tags_;
+    std::vector<std::string> keywords_;
+    std::unique_ptr<bool> is_aggregator_;                 // Optional, hence a pointer
 };
 
 } // namespace internal