+++ include/unity/scopes/internal/zmq_middleware/ZmqMiddleware.h
@@ -18,6 +18,7 @@
 
 #pragma once
 
+#include <unity/scopes/internal/Logger.h>
 #include <unity/scopes/internal/MiddlewareBase.h>
 #include <unity/scopes/internal/MWRegistryProxyFwd.h>
 #include <unity/scopes/internal/MWReplyProxyFwd.h>
@@ -69,6 +70,7 @@ public:
     virtual MWQueryProxy create_query_proxy(std::string const& identity, std::string const& endpoint) override;
     virtual MWQueryCtrlProxy create_query_ctrl_proxy(std::string const& identity, std::string const& endpoint) override;
     virtual MWStateReceiverProxy create_state_receiver_proxy(std::string const& identity) override;
+    virtual MWStateReceiverProxy create_registry_state_receiver_proxy(std::string const& identity) override;
 
     virtual MWQueryCtrlProxy add_query_ctrl_object(QueryCtrlObjectBase::SPtr const& ctrl) override;
     virtual void add_dflt_query_ctrl_object(QueryCtrlObjectBase::SPtr const& ctrl) override;
@@ -130,7 +132,8 @@ private:
     State state_;
     std::condition_variable state_changed_;
     mutable std::mutex state_mutex_;            // Protects state_
-    std::atomic_bool shutdown_flag;
+    std::atomic_bool shutdown_flag_;
+    boost::log::sources::severity_channel_logger_mt<>& logger_;
 
     int64_t twoway_timeout_;                    // Default timeout for twoway invocations
     int64_t locate_timeout_;                    // Timeout for registry locate()