+++ scoperegistry/ScopesWatcher.h
@@ -33,15 +33,17 @@ class ScopesWatcher : public DirWatcher
 {
 public:
     ScopesWatcher(unity::scopes::internal::RegistryObject::SPtr registry,
-                  std::function<void(std::pair<std::string, std::string> const&)> ini_added_callback);
+                  std::function<void(std::pair<std::string, std::string> const&)> ini_added_callback,
+                  boost::log::sources::severity_channel_logger_mt<>& logger);
 
     ~ScopesWatcher();
 
-    void add_install_dir(std::string const& dir, bool notify = false);
+    void add_install_dir(std::string const& dir);
 
 private:
     unity::scopes::internal::RegistryObject::SPtr const registry_;
     std::function<void(std::pair<std::string, std::string> const&)> const ini_added_callback_;
+    boost::log::sources::severity_channel_logger_mt<>& logger_;
     std::map<std::string, std::string> sdir_to_ini_map_;
     std::map<std::string, std::set<std::string>> idir_to_sdirs_map_;
     std::mutex mutex_;
@@ -50,7 +52,7 @@ private:
 
     void remove_install_dir(std::string const& dir);
 
-    void add_scope_dir(std::string const& dir, bool notify);
+    void add_scope_dir(std::string const& dir);
     void remove_scope_dir(std::string const& dir);
 
     void watch_event(DirWatcher::EventType event_type,