+++ src/scopes/internal/ScopeMetadataImpl.cpp
@@ -61,7 +61,7 @@ ScopeMetadataImpl::ScopeMetadataImpl(ScopeMetadataImpl const& other)
     , results_ttl_type_(other.results_ttl_type_)
     , child_scope_ids_(other.child_scope_ids_)
     , version_(other.version_)
-    , tags_(other.tags_)
+    , keywords_(other.keywords_)
 {
     if (other.art_)
     {
@@ -95,6 +95,10 @@ ScopeMetadataImpl::ScopeMetadataImpl(ScopeMetadataImpl const& other)
     {
         location_data_needed_.reset(new bool(*other.location_data_needed_));
     }
+    if (other.is_aggregator_)
+    {
+        is_aggregator_.reset(new bool(*other.is_aggregator_));
+    }
 }
 
 ScopeMetadataImpl& ScopeMetadataImpl::operator=(ScopeMetadataImpl const& rhs)
@@ -119,7 +123,8 @@ ScopeMetadataImpl& ScopeMetadataImpl::operator=(ScopeMetadataImpl const& rhs)
         location_data_needed_.reset(rhs.location_data_needed_ ? new bool(*rhs.location_data_needed_) : nullptr);
         child_scope_ids_ = rhs.child_scope_ids_;
         version_ = rhs.version_;
-        tags_ = rhs.tags_;
+        keywords_ = rhs.keywords_;
+        is_aggregator_.reset(rhs.is_aggregator_ ? new bool(*rhs.is_aggregator_) : nullptr);
     }
     return *this;
 }
@@ -241,9 +246,18 @@ int ScopeMetadataImpl::version() const
     return version_;
 }
 
-std::vector<std::string> ScopeMetadataImpl::tags() const
+std::vector<std::string> ScopeMetadataImpl::keywords() const
 {
-    return tags_;
+    return keywords_;
+}
+
+bool ScopeMetadataImpl::is_aggregator() const
+{
+    if (is_aggregator_)
+    {
+        return *is_aggregator_;
+    }
+    return false;
 }
 
 void ScopeMetadataImpl::set_scope_id(std::string const& scope_id)
@@ -321,6 +335,11 @@ void ScopeMetadataImpl::set_location_data_needed(bool location_data_needed)
     location_data_needed_.reset(new bool(location_data_needed));
 }
 
+void ScopeMetadataImpl::set_is_aggregator(bool is_aggregator)
+{
+    is_aggregator_.reset(new bool(is_aggregator));
+}
+
 void ScopeMetadataImpl::set_child_scope_ids(std::vector<std::string> const& ids)
 {
     child_scope_ids_ = ids;
@@ -335,9 +354,9 @@ void ScopeMetadataImpl::set_version(int v)
     version_ = v;
 }
 
-void ScopeMetadataImpl::set_tags(std::vector<std::string> const& tags)
+void ScopeMetadataImpl::set_keywords(std::vector<std::string> const& keywords)
 {
-    tags_ = tags;
+    keywords_ = keywords;
 }
 
 namespace
@@ -425,14 +444,18 @@ VariantMap ScopeMetadataImpl::serialize() const
         }
         var["child_scopes"] = Variant(va);
     }
-    if (tags_.size())
+    if (keywords_.size())
     {
         VariantArray va;
-        for (auto const& tag: tags_)
+        for (auto const& keyword: keywords_)
         {
-            va.push_back(Variant(tag));
+            va.push_back(Variant(keyword));
         }
-        var["tags"] = Variant(va);
+        var["keywords"] = Variant(va);
+    }
+    if (is_aggregator_)
+    {
+        var["is_aggregator"] = *is_aggregator_;
     }
 
     return var;
@@ -579,14 +602,20 @@ void ScopeMetadataImpl::deserialize(VariantMap const& var)
         }
     }
 
-    tags_.clear();
-    it = var.find("tags");
+    keywords_.clear();
+    it = var.find("keywords");
     if (it != var.end())
     {
         for (auto const& v: it->second.get_array())
         {
-            tags_.push_back(v.get_string());
+            keywords_.push_back(v.get_string());
         }
+    }
+
+    it = var.find("is_aggregator");
+    if (it != var.end())
+    {
+        is_aggregator_.reset(new bool(it->second.get_bool()));
     }
 }