+++ test/gtest/scopes/IdleShutdown/IdleShutdown_test.cpp
@@ -67,10 +67,10 @@ private:
     condition_variable cond_;
 };
 
-void slowsearchscope_thread(RuntimeImpl::SPtr const& rt, string const& runtime_ini_file)
+void slowsearchscope_thread(RuntimeImpl::SPtr const& rt)
 {
     SlowSearchScope scope;
-    rt->run_scope(&scope, runtime_ini_file, "SlowSearchScope.ini");
+    rt->run_scope(&scope, "SlowSearchScope.ini");
 }
 
 // Check that the idle timeout for a server works if no operations are in progress
@@ -84,7 +84,7 @@ TEST(IdleTimeout, server_idle_timeout_while_idle)
 
         // Make a run time for the scope and run the scope.
         auto rt = RuntimeImpl::create("SlowSearchScope", "Runtime.ini");
-        rt->run_scope(&scope, "Runtime.ini", "SlowSearchScope.ini");
+        rt->run_scope(&scope, "SlowSearchScope.ini");
     }
 
     // Check that the scope has indeed timed out. The server shuts down after 2 seconds,
@@ -117,7 +117,7 @@ TEST(IdleTimeout, server_idle_timeout_while_operation_in_progress)
     {
         // Make a run time for the scope and run the scope.
         RuntimeImpl::SPtr srt = RuntimeImpl::create("SlowSearchScope", "Runtime.ini");
-        std::thread slowsearchscope_t(slowsearchscope_thread, srt, "Runtime.ini");
+        std::thread slowsearchscope_t(slowsearchscope_thread, srt);
 
         // Give scope some time to bind to endpoint.
         this_thread::sleep_for(chrono::milliseconds(200));