+++ RELEASE_NOTES.md
@@ -1,6 +1,12 @@
 Release notes
 =============
 
+Changes in version 0.6.10
+=========================
+  - Renamed "Tags" scope .ini option to "Keywords".
+  - Added support for IsAggregator scope .ini option.
+  - Implemented BufferedResultForwarder API in utility namespace.
+
 Changes in version 0.6.9
 ========================
   - Added support for ChildScopes scope .ini option to list scopes ids of aggregated scopes.